backend/requirements.txt
`     2348beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.2
flask-cors==6.0.2
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pillow==12.1.1
PyMuPDF==1.27.2.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
redis==7.3.0
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
spotipy==2.26.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
openpyxl==3.1.5
